fix(archive,cost-center,edefter,efatura): int4 0 as total_count → 0::bigint decode bug (#162)
Single-row queries (find_by_id / find_by_uuid / find_by_invoice_id /
find_by_type / find_active / create RETURNING / update RETURNING / restore
RETURNING) appended `0 as total_count` to satisfy the FromRow structs'
`total_count: Option<i64>` (int8) field. The literal `0` is int4 (OID 23),
which sqlx cannot decode into i64 — so decoding an EXISTENT row failed with
HTTP 500. The non-existent-id case returned no row to decode (→ 404),
hiding the bug; the hurl suite only tested /999999→404.

Fix: cast to `0::bigint` (int8, OID 20) so the literal matches the Row
field. The list queries already use `COUNT(*) OVER() as total_count`
(int8) and are unaffected — this sweep touches only single-row paths.

Sweep: cost_center (7 sites), edefter (3), efatura (6), archive (11) = 27
sites. tax already fixed in #164. grep -rn "0 as total_count" src/ → 0.

Hurl: 25_cost_centers.hurl gains an existent-id happy path
(GET /cost-centers/1 → 200, $.id==1, $.code=="CC01"; id=1 seeded CC01)
that flushes the decode bug. edefter/efatura/archive lists remain 400
from a separate Query-deserialize bug (out of scope); no seeded existent
id exists to fence their by-id path, so they get the code fix only.

+++ turerp/src/domain/archive/postgres_repository.rs
@@ -94,7 +94,7 @@ impl ArchivePolicyRepository for PostgresArchivePolicyRepository {
             VALUES ($1, $2, $3, $4, $5, $6)
             RETURNING "#,
             archive_policy_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(tenant_id)
@@ -115,7 +115,7 @@ impl ArchivePolicyRepository for PostgresArchivePolicyRepository {
             r#"
             SELECT "#,
             archive_policy_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM archive_policies
             WHERE id = $1 AND tenant_id = $2
             "#
@@ -171,7 +171,7 @@ impl ArchivePolicyRepository for PostgresArchivePolicyRepository {
             r#"
             SELECT "#,
             archive_policy_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM archive_policies
             WHERE tenant_id = $1 AND is_active = true
             ORDER BY created_at DESC
@@ -206,7 +206,7 @@ impl ArchivePolicyRepository for PostgresArchivePolicyRepository {
             WHERE id = $6 AND tenant_id = $7
             RETURNING "#,
             archive_policy_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(&update.name)
@@ -332,7 +332,7 @@ impl ArchiveJobRepository for PostgresArchiveJobRepository {
             VALUES ($1, $2, 'Pending', 0, 0)
             RETURNING "#,
             archive_job_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(tenant_id)
@@ -349,7 +349,7 @@ impl ArchiveJobRepository for PostgresArchiveJobRepository {
             r#"
             SELECT "#,
             archive_job_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM archive_jobs
             WHERE id = $1 AND tenant_id = $2
             "#
@@ -461,7 +461,7 @@ impl ArchiveJobRepository for PostgresArchiveJobRepository {
             WHERE id = $5 AND tenant_id = $6
             RETURNING "#,
             archive_job_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(&status_str)
@@ -485,7 +485,7 @@ impl ArchiveJobRepository for PostgresArchiveJobRepository {
             WHERE id = $1 AND tenant_id = $2 AND status = 'Pending'
             RETURNING "#,
             archive_job_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(id)
@@ -596,7 +596,7 @@ impl ArchiveRecordRepository for PostgresArchiveRecordRepository {
             VALUES ($1, $2, $3, $4, $5)
             RETURNING "#,
             archive_record_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(tenant_id)
@@ -616,7 +616,7 @@ impl ArchiveRecordRepository for PostgresArchiveRecordRepository {
             r#"
             SELECT "#,
             archive_record_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM archive_records
             WHERE id = $1 AND tenant_id = $2
             "#
@@ -772,7 +772,7 @@ impl ArchiveRecordRepository for PostgresArchiveRecordRepository {
             WHERE id = $1 AND tenant_id = $2 AND restored_at IS NULL
             RETURNING "#,
             archive_record_columns!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(id)
